requirements.txt
# Core
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
PyYAML>=6.0

# ML
scikit-learn>=1.4.0
xgboost>=2.0.0
shap>=0.45.0

# Deep Learning (PINN)
torch>=2.2.0
torchvision>=0.17.0

# Signal Processing
matplotlib>=3.8.0
seaborn>=0.13.0

# Dashboard
streamlit>=1.35.0
plotly>=5.20.0
streamlit-option-menu>=0.3.12

# Data download
requests>=2.31.0
tqdm>=4.66.0

# Dev / Testing
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.4.0
